fix: move proof pack parsing into agent transform
